+++ src/TodoForm.jsx
@@ -1,0 +1,13 @@
+function TodoForm(){
+    return(
+        <>
+        <form>
+            <label htmlFor="todoTitle">Todo</label>
+            <input id="todoTitle"></input>
+            <button>Add Todo</button>
+        </form>
+        </>
+    )
+}
+
+export default TodoForm
